+++ src/components/input/UInput.ts
@@ -115,6 +115,8 @@ export class UInput extends UFormControlElement<string> {
 
           <input part="input"
             type=${this.type === 'password' && this.showPassword ? 'text' : this.type}
+            aria-label=${ifDefined(this.label)}
+            aria-description=${ifDefined(this.description)}
             name=${ifDefined(this.name)}
             ?required=${this.required}
             ?disabled=${this.disabled}